% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_Names_SQL.R
\name{adding_Names_SQL}
\alias{adding_Names_SQL}
\title{adding names from a similarity table to the 'name' column in a sql
database.}
\usage{
adding_Names_SQL(con_merged, sim_table)
}
\arguments{
\item{con_merged}{'character(1)' connection to the sqlite database.}

\item{sim_table}{'character(1)' path to the similarity table with the matches
of some compounds.

@return fill the 'name' column in the ms_compounds table the names from
similar compounds.

@import DBI
@import dplyr
@import tools

@author Ahlam Mentag

@export}
}
\description{
adding names from a similarity table to the 'name' column in a sql
database.
}
