% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aligning_FTn_FTp_SQL.R
\name{Aligning_FTn_FTp_SQL}
\alias{Aligning_FTn_FTp_SQL}
\title{Aligning compounds from the same instrument}
\usage{
Aligning_FTn_FTp_SQL(
  FTn_path,
  FTp_path,
  FTn_expnr,
  FTp_expnr,
  Assoc = NULL,
  err = 0.02,
  t.ini = 5,
  lc.err = 1,
  rng = 2,
  minIon = 0.01,
  minPeaks = 0.6,
  startpoint = 1,
  save_assoc = FALSE
)
}
\arguments{
\item{FTn_path}{\code{character(1)} the path to the reference database with
negative polarity.}

\item{FTp_path}{\code{character(1)} the path of the SQL database to align with.}

\item{FTn_expnr}{\code{numeric(1)} the reference experiment to align.}

\item{FTp_expnr}{\code{numeric(1)} the experiment to align with the previous
reference experiment.}

\item{Assoc}{\code{character(1)} the file that contains some previous alignment
results done by the user. If \code{NULL} we create it in the function.}

\item{err}{\code{numeric(1)} is the m/z tolerance window used to decide which QTOF
peaks are potential isomeric matches for a given FT peak.}

\item{t.ini}{\code{numeric(1)} used in the \code{Aligning_General_SQL()} to define the
number of neighboring peaks.}

\item{lc.err}{\code{numeric(1)} the retention-time alignment tolerance.}

\item{rng}{\code{numeric(1)} used in \code{RemoveOutliers()} refers to the number of
standard  deviations used to define the outlier cutoff.}

\item{minIon}{\code{numeric(1)}
The minimum matching ions between spectra of peak pairs, which are found
after applying the regression model, and which are the final matching
written in the Assoc table. (default \code{0.01}).}

\item{minPeaks}{Numeric. Minimum MS2 matching ratio required.}

\item{startpoint}{\code{numeric(1)} used in \code{RegressionPie_LCalign()} refers to
the minimum retention time for the first knot in piecewise regression.}

\item{save_assoc}{\code{logical(1)} If true we add the alignment result to the
Assoc text file.}
}
\value{
A \code{data.frame} containing the matched FTMSn-FTMSp or QTOFn-QTOFp
compounds with columns:
\itemize{
\item \code{ref_compid}: FTMSneg or QTOFneg compound ID
\item \code{target_compid}: FTMSpos or QTOFpos compound ID
\item \code{ref_database}: filename of the reference database
\item \code{target_database}: filename of the database to align with.
This table is updated with new associations found by the alignment procedure.
}
}
\description{
\code{Aligning_FTn_FTp_SQL()} can align FTneg-FTpos or QTOFneg-QTOFpos, the
reference experiment is always the one with the negative ionization mode
and adds the aligned compounds to a given text file.
}
\details{
the function reads the both sql databases FTMS or QTOF, and extracts
their modes, create an output file to store the alignment results in a .txt
file if it's already given by the user otherwise we create a new one.
Then calls a set of defined functions :
\itemize{
\item \code{\link[=Aligning_General_SQL]{Aligning_General_SQL()}}: performs the local alignment, and adjust the mz
values if the alignment is between two different polarities.
\item \code{matchFTSyn_SQL()}: removes the matches found by the previous function
if they contain fewer than \code{minIon} matched peaks.
\item \code{RemoveOutliers()}: calculate the standard deviation in the rt difference
found between each matched feature and remove all rows outside a range.
\item \code{RegressionPie_LCalign()}: apply a regression model  to reduce the
influence of false matches and outliers.
\item \code{FillAssocFTnQTOFn_SQL()}: store the final aligned compounds to a .txt
file.
}
}
\author{
Ahlam Mentag
}
