% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aligning_General_SQL.R
\name{Aligning_General_SQL}
\alias{Aligning_General_SQL}
\title{Local alignment between FTMS and QTOF datasets}
\usage{
Aligning_General_SQL(FT_con, QTOF_con, FT_expnr, QTOF_expnr, err, t.ini)
}
\arguments{
\item{FT_con}{\code{character(1)} A DBI database connection object to the FTMS
SQL database.}

\item{QTOF_con}{\code{character(1)} A DBI database connection object to the QTOF
SQL database.}

\item{FT_expnr}{\code{numeric(1)} Experiment number (expid) to load from the FTMS
SQL database.}

\item{QTOF_expnr}{\code{numeric(1)} Experiment number to load from the QTOF SQL
database.}

\item{err}{\code{numeric(1)} is the m/z tolerance window used to decide which QTOF
peaks are potential isomeric matches for a given FT peak.}

\item{t.ini}{\code{numeric(1)}The number of neighboring peaks}
}
\value{
A data.frame representing the locally aligned features
}
\description{
The \code{Aligning_General_SQL()} function performs local alignment between two
LC-MS datasets stored in SQL databases (FTMS and QTOF).
The function:
\itemize{
\item extracts compound data from the both SQL databases,
\item performs local alignment using \code{LCalign()},
\item adjusts \code{mass_measured} values when the two experiments have different
polarities.
}
}
\author{
Ahlam Mentag
}
