% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCSPP_SQL.R
\name{allCSPP_SQL}
\alias{allCSPP_SQL}
\title{generating compound pairs with their associated similarity metrics.}
\usage{
allCSPP_SQL(sql_path, exp.id, nr_col = NULL, thr1 = 1, thr2 = 0.9, thr3 = 0.4)
}
\arguments{
\item{sql_path}{'character(1)' path to the sqlite database.}

\item{exp.id}{'number(1)' experiment to use for network generation.}

\item{nr_col}{number of columns in the compound_add table.}

\item{thr1}{'number(1)' minimum number of product ions (varying between 0
and 1) of one CID spectrum that can be traced in another CID spectrum,
set on 1 by default.}

\item{thr2}{'number(1)' dot product threshold for the common ions
between two CID spectra, set on 0.9 by default.}

\item{thr3}{'number(1)' average of the dot products obtained for the common
product ions and the common neutral losses, by default set on 0.4.}
}
\value{
a data frame of compound pairs with their associated CSPP similarity
metrics.
}
\description{
allCSPP_SQL() connects to a SQLite database, retrieves CSPP-based
relationships between compounds for a given experiment (exp.id), parses
pairwise similarity information from the gnps_add table, filters the results
using defined thresholds (thr1, thr2, thr3).
}
\author{
Ahlam Mentag
}
