% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutomFeatureGroup_SQL.R
\name{AutomFeatureGroup_SQL}
\alias{AutomFeatureGroup_SQL}
\title{Grouping of the features.}
\usage{
AutomFeatureGroup_SQL(dat, rtmed, ints, stp, y10, y20, fc)
}
\arguments{
\item{dat}{the output of the featureDefinitions() of the xcmsExperiment
object.}

\item{rtmed}{the median retention times extracted from the
featureDefinitions() function.}

\item{ints}{the intensities of the features in each sample.}

\item{stp}{set within the XCMSgrouping() function as the output from the
BreakRowsXCMS() function: represents the indices of features with a
particular retention time on which to split.}

\item{y10}{maximum retention time deviation for feature alignment across
chromatograms, 1 sec by default.}

\item{y20}{minimum feature abundance-based Pearson correlation for feature
alignment across chromatograms, 0.8 by default.}

\item{fc}{set within the XCMSgrouping() function as the output from the
LoadProc() function: represents the genotype/treatment class information.}

\item{tR}{(e.g. 1 sec) defines the retention time difference between two
subsequent rows that is needed to define a breaking row}
}
\value{
an XCMSExperiment object with the new grouping columns.
}
\description{
the \code{AutomFeatureGroup_SQL()} function sort the feature information by
retention time and then calls the featureGroupsXcms() function to perform
the grouping instructions inside it.
}
\author{
Ahlam Mentag
}
