% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakRowsXCMS_SQL.R
\name{BreakRowsXCMS_SQL}
\alias{BreakRowsXCMS_SQL}
\title{Breaking featureDefinitions() rows.}
\usage{
BreakRowsXCMS_SQL(dat, tR, rtmed)
}
\arguments{
\item{dat}{the output of the featureDefinitions() of the xcmsExperiment
object}

\item{tR}{(e.g. 1 sec) defines the retention time difference between two
subsequent rows that is needed to define a breaking row}

\item{rtmed}{the median retention times extracted from the
featureDefinitions() function.}
}
\value{
a list of indices of each breaking row.
}
\description{
creates list of breaking rows to enable the subsequent peakgrouping of a
xcmsExperiment object.
}
\author{
Ahlam Mentag
}
