% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_compid_assoc.R
\name{convert_compid_assoc}
\alias{convert_compid_assoc}
\title{This script converts the \emph{Assoc} text files from the original to the new
\emph{long} format which is more flexible and expandable.}
\usage{
convert_compid_assoc(input_path, output_path)
}
\description{
This script converts the \emph{Assoc} text files from the original to the new
\emph{long} format which is more flexible and expandable.
}
\author{
Ahlam Mentag
}
