% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertXCMS4_SQL.R
\name{ConvertXCMS4}
\alias{ConvertXCMS4}
\title{Performing a re-indexing and re-ordering of feature groups based
on retention time.}
\usage{
ConvertXCMS4(XCMS4)
}
\arguments{
\item{XCMS4}{the resulting \code{XcmsExperiment} object from the
\code{AutomFindingISF_SQL()} function.}
}
\value{
An updated \code{XcmsExperiment} object with new columns in
featureDefinitions, "MeAb" and "CorrPKGrp":

\if{html}{\out{<div class="sourceCode">}}\preformatted{    - `MeAb` : Mean retention time (rtmed) of the feature’s peak group 
             (CON.new).
      
    - `massMS1` : Corrected sequential peak-group index after RT-based 
                  reordering.
}\if{html}{\out{</div>}}
}
\description{
The \code{ConvertXCMS4_SQL()} computes the mean retention time for each
feature group, orders features accordingly, and generates a corrected,
sequential peak-group index for downstream ion–neutral and combinatorial
analyses.
}
\author{
Ahlam Mentag
}
