% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspp.tot_SQL.R
\name{cspp.tot_SQL}
\alias{cspp.tot_SQL}
\title{Compute CSPP scores for one experiment and update the compound_add SQL
table}
\usage{
cspp.tot_SQL(
  sql_path,
  expid,
  mzerr = 0.015,
  cspp = "cspp.txt",
  peakwidth = NULL,
  IntThres = 100
)
}
\arguments{
\item{sql_path}{\code{Character(1)} string giving the path to the SQLite database.}

\item{expid}{\code{Integer(1)} experiment id. Only compounds belonging to this
experiment are processed.}

\item{mzerr}{\code{Numeric(1)} mass tolerance (in Da) used to match precursor and
product ions. Default it is configured for ftms data
\code{0.01}, for qtof data the user could set it to \code{0.015}.}

\item{cspp}{\code{Character(1)} string giving the path to the CSPP configuration
file. This file defines the conversion types, mass differences,
elution order, and the target columns in \code{compound_add}.}

\item{peakwidth}{\code{Numeric(1)} retention time window used to enforce elution
order constraints between substrate and product compounds. If \code{NULL},
defaults to \code{0.2}.}

\item{IntThres}{Numeric intensity threshold applied to MS/MS fragment ions
before similarity calculations. Default it is configured for ftms data
\code{100}, for qtof data the user could set it to \code{5}.}
}
\value{
Invisibly returns the updated \code{compound_add} data frame
corresponding to the processed experiment. The primary effect of the
function is the update of the SQL table.
}
\description{
cspp.tot_SQL() computes CSPP (Conversion-Specific Product Pair) scores for all
compounds belonging to a given experiment and updates the corresponding rows
in the \code{compound_add} table of a SQLite database.
}
\details{
For each compound in the experiment, theoretical conversion products are
generated based on the conversion definitions provided in a CSPP configuration
file. Candidate precursor/product pairs are matched using mass differences
and retention time constraints, and their MS/MS spectra are compared to
compute similarity scores.

No rows representing the different compound IDs of the experiment need to be
pre-created in the \code{compound_add} table. During execution, CSPP scores
are written to the columns specified in the CSPP configuration file. If no
valid conversions are found for a given compound, the corresponding CSPP
fields are left as \code{NA}.
}
\author{
Ahlam Mentag
}
