% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSgrouping2_SQL.R, R/XCMSgrouping_SQL.R
\name{.create_name}
\alias{.create_name}
\title{helper function to a create a name for the features.}
\usage{
.create_name(mzmed, rtmed)

.create_name(mzmed, rtmed)
}
\description{
helper function to a create a name for the features.

helper function to a create a name for the features.
}
