% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureAnnotXCMS_SQL.R
\name{FeatureAnnotXCMS_SQL}
\alias{FeatureAnnotXCMS_SQL}
\title{Annotate peaks in XCMS feature groups}
\usage{
FeatureAnnotXCMS_SQL(XCMS, y11, fc, ints)
}
\arguments{
\item{XCMS}{an \code{XCMSExperiment} object with the previous grouping columns}

\item{y11}{monoisotopic mass of the buffer adduct, 46 Da (formic acid)
by default.}

\item{fc}{set within the XCMSgrouping() function as the output from the
LoadProc() function: represents the genotype/treatment class information.}

\item{ints}{the intensities of the features in each sample.}
}
\value{
an XCMSExperiment object with the new annotation columns.
}
\description{
the \code{FeatureAnnotXCMS_SQL()} performs the annotation of the features within
each FG as 13C isotopes ‘anno’ column , buffer adducts ‘anno.1’ column
or ion/neutral complex between the molecule and its ion ‘anno.2’ column.
}
\author{
Ahlam Mentag
}
