% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureGroupXCMS_SQL.R
\name{FeatureGroupXCMS_SQL}
\alias{FeatureGroupXCMS_SQL}
\title{Grouping of the features.}
\usage{
FeatureGroupXCMS_SQL(dt, ints, y10, y20, fc)
}
\arguments{
\item{dt}{the output of the featureDefinitions() of the xcmsExperiment
object ordred by retention time.}

\item{ints}{the intensities of the features in each sample.}

\item{y10}{maximum retention time deviation for feature alignment across
chromatograms, 1 sec by default.}

\item{y20}{minimum feature abundance-based Pearson correlation for feature
alignment across chromatograms, 0.8 by default.}

\item{fc}{set within the XCMSgrouping() function as the output from the
LoadProc() function: represents the genotype/treatment class information.}

\item{tR}{(e.g. 1 sec) defines the retention time difference between two
subsequent rows that is needed to define a breaking row}
}
\value{
an XCMSExperiment object with the new grouping columns.
}
\description{
the \code{FeatureGroupXCMS_SQL()} function performs the m/z feature grouping
based on equal retention times and a high Pearson correlation across
biological replicates.
}
\author{
Ahlam Mentag
}
