% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSubgroupXCMS_SQL.R
\name{FeatureSubgroupXCMS_SQL}
\alias{FeatureSubgroupXCMS_SQL}
\title{Further grouping of the features.}
\usage{
FeatureSubgroupXCMS_SQL(XCMS, y10)
}
\arguments{
\item{XCMS}{an \code{XCMSExperiment} object with the previous grouping columns}

\item{y10}{maximum retention time deviation for feature alignment across
chromatograms, 1 sec by default.}
}
\value{
an XCMSExperiment object with the new grouping columns.
}
\description{
the \code{FeatureSubgroupXCMS_SQL()} function performs further grouping based
on the median retention time of each current group
}
\author{
Ahlam Mentag
}
