% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Files_TransferNames_SQL.R
\name{Files_TransferNames_SQL}
\alias{Files_TransferNames_SQL}
\title{Load compound metadata from a SQLite database and association file}
\usage{
Files_TransferNames_SQL(sqlite_dir, assoc_path)
}
\arguments{
\item{sqlite_dir}{Character string. Directory containing SQLite databases.
Each database file must be named <database_name>.sqlite.}

\item{assoc_path}{Character string. Path to the association file defining
reference and target databases and compound ID mappings.}
}
\value{
A list with four elements:
\describe{
\item{ftng}{Data frame of compounds from the reference database.}
\item{ftps}{Data frame of compounds from the target database.}
\item{syn}{Data frame containing the union of reference and target compounds.}
\item{Assoc}{The association data frame as read from \code{assoc_path}.}
}
}
\description{
Files_TransferNames_SQL() function
reads compound metadata from a SQLite database and prepares
subdatabase specific compound tables based on the database names defined
in a compound association file.
}
\author{
Ahlam Mentag
}
