% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillAssocFTnFTpn_SQL.R
\name{FillAssocFTnFTpn_SQL}
\alias{FillAssocFTnFTpn_SQL}
\title{Associate FT and QTOF features using SQL and MS2 matching}
\usage{
FillAssocFTnFTpn_SQL(
  FTn_con,
  FTp_con,
  Assoc,
  FTn_expnr,
  FTp_expnr,
  cutoff = 1,
  rg = rg,
  lc.err,
  err,
  minIon,
  polarity_ftn = 0,
  polarity_ftp = 1,
  FTn_path,
  FTp_path
)
}
\arguments{
\item{FTn_con}{\code{DBIConnection}
A DBI connection object to the reference SQLite database.}

\item{FTp_con}{\code{DBIConnection}
A DBI connection object to the SQLite database to align with.}

\item{Assoc}{\code{data.frame} A table storing previously detected associations.
Will be filled with the new association if they are not already
in the assoc table.}

\item{FTn_expnr}{\code{numeric(1)} the reference experiment to align.}

\item{FTp_expnr}{\code{numeric(1)} the experiment to align with the previous
reference experiment.}

\item{cutoff}{\code{numeric(1)}
Minimum retention time to consider for FT compounds.}

\item{rg}{\code{numeric(6)}
Regression coefficients for retention-time alignment.}

\item{lc.err}{\code{numeric(1)}
Allowed retention-time error window after regression transformation.}

\item{err}{\code{numeric(1)}
Error threshold for candidate match search (\code{Find_cand_matches_SQL}).}

\item{minIon}{\code{numeric(1)}
The minimum matching ions between spectra of peak pairs, which are found
after applying the regression model, and which are the final matching
written in the Assoc table. (default \code{0.02}).}

\item{polarity_ftn}{\code{Integer (0/1)} Polarity of negative FTMS or QTOF experiment.}

\item{polarity_ftp}{\code{Integer (0/1)} Polarity of positive FTMS or QTOF experiment.}

\item{FTn_path}{\code{character(1)} the path to the reference database with
negative polarity.}

\item{FTp_path}{\code{character(1)} the path of the SQL database to align with.}

\item{aggregated_Ft}{\code{logical(1)}
Whether to use aggregated FTMS spectra.}

\item{aggregated_QTOF}{\code{logical(1)}
Whether to use aggregated QTOF MS2 spectra.}
}
\value{
\code{data.frame}
The updated association table with new FTMS–QTOF matches appended.
}
\description{
\code{FillAssocFTnFTpn_SQL()} finds associations between experiments from the same
instrument type, the one with negative polarity should be considered as
reference, based on retention time alignment, MS2 peak similarity,
and alignment rules. It updates and returns the association table (\code{Assoc}).
}
\details{
The function:
\itemize{
\item loads both databases compounds from SQL databases,
\item computes expected RT alignment using regression parameters,
\item finds candidate matches using SQL-assisted search,
\item filters candidates using MS2 peak similarity,
\item appends validated associations to \code{Assoc}.
}
}
\author{
Ahlam Mentag
}
