% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillAssocFTnQTOFn_SQL.R
\name{FillAssocFTnQTOFn_SQL}
\alias{FillAssocFTnQTOFn_SQL}
\title{Associate FT and QTOF features using SQL and MS2 matching}
\usage{
FillAssocFTnQTOFn_SQL(
  FT_con,
  QTOF_con,
  Assoc,
  FT_expnr,
  QTOF_expnr,
  cutoff,
  rg,
  lc.err,
  err,
  minIon = 0.02,
  polarity_ft = 0,
  polarity_qtof = 0,
  FT_path,
  QTOF_path,
  aggregated_Ft = FALSE,
  aggregated_QTOF = FALSE
)
}
\arguments{
\item{FT_con}{\code{DBIConnection}
A DBI connection object to the FT SQLite database.}

\item{QTOF_con}{\code{DBIConnection}
A DBI connection object to the QTOF SQLite database.}

\item{Assoc}{\code{data.frame} A table storing previously detected associations.
Will be filled with the new association if they are not already
in the assoc table.}

\item{FT_expnr}{\code{numeric(1)}
The reference experiment number to load from the FTMS SQL database.}

\item{QTOF_expnr}{\code{numeric(1)}
Experiment number to load from the QTOF SQL database.}

\item{cutoff}{\code{numeric(1)}
Minimum retention time to consider for FT compounds.}

\item{rg}{\code{numeric(6)}
Regression coefficients for retention-time alignment.}

\item{lc.err}{\code{numeric(1)}
Allowed retention-time error window after regression transformation.}

\item{err}{\code{numeric(1)}
Error threshold for candidate match search (\code{Find_cand_matches_SQL}).}

\item{minIon}{\code{numeric(1)}
The minimum matching ions between spectra of peak pairs, which are found
after applying the regression model, and which are the final matching
written in the Assoc table. (default \code{0.02}).}

\item{polarity_ft}{\code{integer(1)}
Polarity filter for FTMS MS2 spectra (0 or 1).}

\item{polarity_qtof}{\code{integer(1)}
Polarity filter for QTOF MS2 spectra (0 or 1).}

\item{FT_path}{\code{character(1)}
File path to the FTMS database .}

\item{QTOF_path}{\code{character(1)}
File path to the QTOF database.}

\item{aggregated_Ft}{\code{logical(1)}
Whether to use aggregated FTMS spectra.}

\item{aggregated_QTOF}{\code{logical(1)}
Whether to use aggregated QTOF MS2 spectra.}
}
\value{
\code{data.frame}
The updated association table with new FTMS–QTOF matches appended.
}
\description{
\code{FillAssocFTnQTOFn_SQL()} finds associations between FTMS and QTOFMS
compounds based on retention time alignment, MS2 peak similarity, and
alignment rules. It updates and returns the association table (\code{Assoc}).
}
\details{
The function:
\itemize{
\item loads FT and QTOF compounds from SQL databases,
\item computes expected RT alignment using regression parameters,
\item finds candidate matches using SQL-assisted search,
\item filters candidates using MS2 peak similarity,
\item appends validated associations to \code{Assoc}.
}
}
\author{
Ahlam Mentag
}
