% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingISF_SQL.R
\name{FindingISF_SQL}
\alias{FindingISF_SQL}
\title{finding in-source fragments.}
\usage{
FindingISF_SQL(fg, XCMS, spd, ms2.ft)
}
\arguments{
\item{fg}{set within the AutomFindingISF() function, represents the feature
group.}

\item{XCMS}{the resulting \code{XcmsExperiment} object from the
\code{XCMSgrouping_SQL()} function.}

\item{spd}{spectra Data of the \code{XcmsExperiment} object.}

\item{ms2.ft}{peaks data of the \code{XcmsExperiment} object.}
}
\value{
An updated \code{XcmsExperiment} object with new annotation column
created inside FindingISF_SQL() function.
}
\description{
finding whether a m/z feature is an in-source fragment (ISF) of another m/z
feature in a particular feature group.
fg is the feature group number.
}
\author{
Ahlam Mentag
}
