% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignedCompids.R, R/getAlignedCompids_SQL.R
\name{getAlignedCompids_SQL}
\alias{getAlignedCompids_SQL}
\title{Extracting the aligned compound IDs with their database names of given
compounds.}
\usage{
getAlignedCompids_SQL(compound_id, database_name, alignment_file)

getAlignedCompids_SQL(compound_id, database_name, alignment_file)
}
\arguments{
\item{compound_id}{'vector()' given compound_id, single value a vector of
compounds, to extract their aligned compounds.}

\item{database_name}{'character(1)' database name of the given compound_id.}

\item{alignment_file}{'character(1)' the file path that contains the
alignment results.}
}
\value{
a 'data.frame' with compid and database.

a 'data.frame' with compid and database.
}
\description{
Extracting the aligned compound IDs with their database names of given
compounds.

Extracting the aligned compound IDs with their database names of given
compounds.
}
\author{
Ahlam Mentag
}
