% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnps.display_SQL.R
\name{gnps.display_SQL}
\alias{gnps.display_SQL}
\title{Display GNPS similarity results from an SQLite database.}
\usage{
gnps.display_SQL(sql_path, dbkey, nr_col2 = 6)
}
\arguments{
\item{sql_path}{\code{Character(1)} string giving the path to the SQLite database.}

\item{dbkey}{'numeric(1)' the compound_id used to select the corresponding
row in the \code{compound_add} table.}

\item{nr_col}{'numeric(1)' maximum number of columns to scan in the table.
Default is 6.}
}
\value{
A \code{data.frame} containing GNPS similarity results with the columns:
\describe{
\item{massdiff}{Mass difference associated with the match.}
\item{compid.prod}{Product compound identifier.}
\item{ions.prod}{Number of product ions.}
\item{ave.common}{Average number of common ions.}
\item{ave.dot}{Average dot-product similarity score.}
}

Returns an empty \code{data.frame} if no matching \code{compound_id} is found.
}
\description{
Extracts and formats GNPS-based similarity information from the
\code{gnps_add} table in an SQLite database for a given \code{compound_id}.
Encoded similarity values stored in the table columns are parsed
and returned as a structured data frame.
}
\author{
Ahlam Mentag
}
