% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadProc_SQL.R
\name{LoadProc_SQL}
\alias{LoadProc_SQL}
\title{featureDefinitions Extraction and sample numbers to a list.}
\usage{
LoadProc_SQL(x)
}
\arguments{
\item{x}{\code{XcmsExperiment} object with the results from xcms preprocessing.}
}
\value{
a list of featureDefinitions() and fc (sample number).
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ 'LoadProc_SQL()' function creates list of featureDefinitions and sample
  numbers. 
}\if{html}{\out{</div>}}
}
\author{
Ahlam Mentag
}
