% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local.net_SQL.R
\name{Local.net_SQL}
\alias{Local.net_SQL}
\title{display the network graph}
\usage{
Local.net_SQL(net.lst, dbkey, nettype, nr_of_seq = 2)
}
\arguments{
\item{net.lst}{'list()' list of nodes and edges.}

\item{dbkey}{'number(1)' refers to the compound_id.}

\item{nettype}{'character(1)' the graph type the user want to display,
it could be either}

\item{nr_of_seq}{'number(1)' size of local network, i.e., number of
subsequent edges starting from the node representing the selected COMPID.}
}
\value{
list of edges and nodes belonging to the local network of the
selected COMPID and plot the network.
}
\description{
The Local.net_SQL() function will select those nodes and edges
belonging to the local network of the selected COMPID and display
the network graph.
}
\author{
Ahlam Mentag
}
