% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFTSyn_sql.R
\name{matchFTSyn_SQL}
\alias{matchFTSyn_SQL}
\title{Match FT and QTOF MS2 spectra based on local alignments}
\usage{
matchFTSyn_SQL(
  LCal,
  FT_con,
  QTOF_con,
  polarity_ft = 0,
  polarity_qtof = 0,
  minPeaks = 0.6,
  aggregated_Ft = FALSE,
  aggregated_QTOF = FALSE
)
}
\arguments{
\item{LCal}{\code{data.frame} A data.frame returned by \code{Aligning_General_SQL()} function.}

\item{FT_con}{\code{character(1)} A DBI connection object
to the FTMS SQLite database (the reference database).}

\item{QTOF_con}{\code{character(1)} A DBI connection object to the QTOF
SQLite database.}

\item{polarity_ft}{\code{numeric(1)} Integer (0 or 1), refers to the polarity
of the ftms  experiment to align.}

\item{polarity_qtof}{\code{numeric(1)}Integer (0 or 1), refers to the polarity
of the qtof experiment to align.}

\item{minPeaks}{\code{Numeric(1)} Minimum MS2 matching ratio required.}

\item{aggregated_Ft}{\code{logical(1)} Use aggregated FTMS spectra.}

\item{aggregated_QTOF}{\code{logical(1)} Use aggregated QTOFMS spectra.}
}
\value{
A filtered version of \code{LCal}, keeping only MS2 supported matches.
}
\description{
\code{matchFTSyn_SQL()} filters locally aligned FTMS–QTOF candidate pairs in
(\code{LCal}) by comparing MS2 peak lists stored in SQL databases.
}
\details{
The function:
\itemize{
\item Loads FT and QTOF MS2 spectra from SQLite via DBI
\item Applies polarity and spectrum-type filtering
\item Computes MS2 peak overlap for each FT–QTOF candidate pair
\item Removes pairs with less matched peaks based on a threshold
}
}
\author{
Ahlam Mentag
}
