% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1specplot_SQL.R
\name{MS1specplot_SQL}
\alias{MS1specplot_SQL}
\title{plot MS1 spectra}
\usage{
MS1specplot_SQL(sql_path, dbkey, lc.err = 0.02, mz.err = 0.001, prcx = 0.7)
}
\description{
plot MS1 spectra
}
\author{
Ahlam Mentag
}
