% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS3specplot_SQL.R
\name{MS3specplot_SQL}
\alias{MS3specplot_SQL}
\title{plot MS3 spectra}
\usage{
MS3specplot_SQL(sql_path, dbkey, prcx = NULL, wh = NULL, single = NULL)
}
\description{
plot MS3 spectra
}
\author{
Ahlam Mentag
}
