% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMSplot_SQL.R
\name{MSMSplot_SQL}
\alias{MSMSplot_SQL}
\title{plot MS2 spectra}
\usage{
MSMSplot_SQL(
  sql_path,
  dbkey,
  prdion,
  neutloss,
  err = NULL,
  minum = NULL,
  oldpar = NULL,
  nl = NULL
)
}
\description{
plot MS2 spectra
}
\author{
Ahlam Mentag
}
