% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnspecplot_SQL.R
\name{MSnspecplot_SQL}
\alias{MSnspecplot_SQL}
\title{plot MSnspectra}
\usage{
MSnspecplot_SQL(
  sql_path,
  dbkey,
  nr_col = 35,
  nr_col2 = 6,
  lc.err = 0.02,
  mz.err = 0.001,
  MS1 = NULL,
  prcx = NULL
)
}
\description{
plot MSnspectra
}
\author{
Ahlam Mentag
}
