% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultExp_GNPS_SQL.R
\name{MultExp_GNPS_SQL}
\alias{MultExp_GNPS_SQL}
\title{feature pairs generation based on highly similar CID spectra.}
\usage{
MultExp_GNPS_SQL(
  sql_path,
  startexp,
  stopexp,
  peakwidth = NULL,
  mzerr = 0.015,
  min = 5,
  adduct = NULL,
  thr1 = NULL,
  thr2 = NULL,
  thr3 = NULL,
  IntThres = 5
)
}
\arguments{
\item{sql_path}{\code{Character(1)} string giving the path to the SQLite database.}

\item{startexp}{\code{Integer(1)} the start experiment id.}

\item{stopexp}{\code{Integer(1)} the last experiment id, where the process should
stop.}

\item{peakwidth}{\code{Numeric(1)} half of the retention time window that is
defined for the CSPP ‘substrate’ feature allowing
a minimum retention time difference with the CSPP ‘product’ feature,
by default set on \code{0.2}.}

\item{mzerr}{\code{Numeric(1)} mass tolerance (in Da) used to match precursor and
product ions. Default it is configured for ftms data
\code{0.01}, for qtof data the user could set it to \code{0.015}.}

\item{min}{minimum absolute intensity of product ion to be withheld for CID
spectral matching, by default \code{5} for QTOF, but the user can set it to
\code{100} for FTMS spectra.}

\item{adduct}{mass of buffer, by default \code{46.0055 Da} representing
formic acid.}

\item{thr1}{average of the number of common ions and the number of common
losses, set on \code{3} by default.}

\item{thr2}{ratio of thr1 to the number of product ions in the CID spectrum
of the feature in the feature pair having the lowest number of product ions,
by default set on \code{0.1}.}

\item{thr3}{average of the dot products obtained for the common product ions
and the common neutral losses, by default set on \code{0.4}.}

\item{IntThres}{Numeric intensity threshold applied to MS/MS fragment ions
before similarity calculations. Default it is configured for ftms data
\code{100}, for qtof data the user could set it to \code{5}.}
}
\value{
Invisibly returns \code{TRUE}. The primary effect of the
function is the update of the \code{gnps_add} SQL table.
}
\description{
MultExp_GNPS_SQL() function generates feature pairs based on highly similar
CID spectra followed by mass difference computation. This method is similar
to the method used by the Global Natural Products Social Molecular Networking
(GNPS) website (Wang et al., 2016). GNPS-like associations are computed and
stored within the ‘gnps_add’ table in the sql database.
}
\author{
Ahlam Mentag
}
