% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.addit_SQL.R
\name{net.addit_SQL}
\alias{net.addit_SQL}
\title{generating the nodes and edges for the experiment-based global network.}
\usage{
net.addit_SQL(
  sql_path,
  exp.id,
  nettype,
  nr_col = NULL,
  min = NULL,
  thr1 = 3,
  thr2 = 0.1,
  thr3 = 0.2
)
}
\arguments{
\item{sql_path}{'character(1)' path to the sqlite database.}

\item{exp.id}{'number(1)' experiment to use for network generation.}

\item{nettype}{'character(1)' the graph type the user want to display,
it could be either}

\item{nr_col}{number of columns in the compound_add and gnps_add tables.}

\item{min}{'numeric(1)' minimum intensity for the spectra.}
}
\value{
list of edges and nodes.
}
\description{
The net.addit_SQL() function is responsible for generating lists of network
nodes and edges by the net.nodes_SQL() and either the allCSPP_SQL()
or allGNPS_SQL() functions, respectively. This generates the nodes and edges
for the experiment-based global network. Only edges are retained that pass
certain CID spectral similarity thresholds.
}
\author{
Ahlam Mentag
}
