% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.nodes_SQL.R
\name{net.nodes_SQL}
\alias{net.nodes_SQL}
\title{generating lists of network nodes and edges.}
\usage{
net.nodes_SQL(sql_path, exp.id, min)
}
\arguments{
\item{sql_path}{'character(1)' path to the sqlite database.}

\item{exp.id}{'number(1)' experiment to use for network generation.}

\item{min}{'numeric(1)' minimum intensity for the spectra.}
}
\description{
net.nodes_SQL() connects to a SQLite database and returns a table of
compounds 'compound_id' for a given experiment 'exp.id', along with
the number of MS2 peaks 'nr_peaks' per compound whose intensity is greater
than or equal to a specified threshold 'min'.
}
\author{
Ahlam Mentag
}
