% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultMatchCID_Spectra.R, R/round_perl.R,
%   R/similarity_functions.R
\name{round_perl}
\alias{round_perl}
\title{Round a number using Perl-style rounding}
\usage{
round_perl(number)

round_perl(number)

round_perl(number)
}
\arguments{
\item{number}{A numeric value to be rounded.}
}
\value{
a rounded mz value.

An integer rounded from \code{number}.

a rounded mz value.
}
\description{
Helper function to round the mz values depending on the resolution type
chosen by the user.

This function rounds a numeric value to the nearest integer using
Perl-style rounding (i.e., adding 0.5 and taking the floor).

Helper function to round the mz values depending on the resolution type
chosen by the user.
}
\examples{
round_perl(2.3)
round_perl(2.7)
}
\author{
Rebecca Dauwe
}
