% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchCombinat_SQL.R
\name{SearchCombinat_SQL}
\alias{SearchCombinat_SQL}
\title{Searching for features that represent combinations of other features.}
\usage{
SearchCombinat_SQL(XCMS4, adjPeak, mode)
}
\arguments{
\item{XCMS4}{the resulting \code{XcmsExperiment} object from the \code{ConvertXCMS4()}
function.}

\item{adjPeak}{\code{numeric()} number of adjacent feature groups in the
chromatogram to search for possible combinations of features, 1 by default.}

\item{mode}{\code{character(1)} is the mode of the experiment, either neg or pos.}
}
\value{
An updated \code{XcmsExperiment} object with new columns in
featureDefinitions, "Comb" and "massMS1":

\if{html}{\out{<div class="sourceCode">}}\preformatted{    - `Comb` : Ion/neutral combination annotation indicating pairs of
      features whose masses sum to another observed feature (formatted
      as `"mz1/mz2"`).
      
    - `massMS1` : Integer-encoded MS1 mass (m/z adjusted for protonation/
       deprotonation and scaled by 100) used for ion–neutral combination
       matching.
}\if{html}{\out{</div>}}
}
\description{
The \code{SearchCombinat_SQL()} function searches for features that represent
combinations of other features present in the proper and in the neighboring
FGs.
}
\author{
Ahlam Mentag
}
