% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSgrouping2_SQL.R
\name{XCMSgrouping2_SQL}
\alias{XCMSgrouping2_SQL}
\title{searching for combination and in-source fragments within the FGs
using CID spectral data.}
\usage{
XCMSgrouping2_SQL(x, adjPeak = 1, mode = NULL)
}
\arguments{
\item{x}{\code{XcmsExperiment} object resulting from the XCMSgrouping() function.}

\item{adjPeak}{\code{numeric()} number of adjacent feature groups in the
chromatogram to search for possible combinations of features, 1 by default.}

\item{mode}{\code{character(1)} is the mode of the experiment, either neg or pos.}
}
\value{
An update \code{XcmsExperiment} object with new and updated columns in
featureDefinitions .
}
\description{
The \code{XCMSgrouping2_SQL()} function searches for in-source fragments within
the FGs using CID spectral data via the nested AutomFindingISF()
function In addition, the function searches for features that represent
combinations of other features present in the proper and in the neighboring
FGs via the SearchCombinat() function. The function takes the xcms resulting
object from the XCMSgrouping() function as input and creates an upgraded
xcms object.
}
\author{
Ahlam Mentag
}
