% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSgrouping_SQL.R
\name{XCMSgrouping_SQL}
\alias{XCMSgrouping_SQL}
\title{Grouping and annotating the features of an \code{XcmsExperiment} object.}
\usage{
XCMSgrouping_SQL(x, tR = 1, y10 = 1, y20 = 0.8, y11 = 46)
}
\arguments{
\item{x}{\code{XcmsExperiment} object with the results from xcms preprocessing}

\item{tR}{\code{numeric(1)} minimum break-up retention time difference: determines
the minimum difference in retention time between two subsequent rows
(i.e., features) prior to considering the row as a candidate for splitting
the ‘featureDefinitions()’-derived spreadsheet (\link{BreakRowsXCMS_SQL}),
1 sec by default.}

\item{y10}{maximum retention time deviation for feature alignment across
chromatograms, 1 sec by default.}

\item{y20}{minimum feature abundance-based Pearson correlation for feature
alignment across chromatograms, 0.8 by default}

\item{y11}{monoisotopic mass of the buffer adduct, 46 Da (formic acid)
by default.}
}
\value{
An update \code{XcmsExperiment} object with new columns in
featureDefinitions of the grouping and annotation information.
}
\description{
The \code{XCMSgrouping_SQL()} takes as input a preprocessed \code{XcmsExperiment}
object, then extract the featureDefinitions for mzmed, rtmed, and name columns
also we call featureValues to get the individual intensities of the features
in each sample columns from it. Hence we add an fc column for the sample
groups. Then we call a set of functions to perform the grouping and
the annotation of the features, and add those information as additional
columns to featureDefinitions.
}
\author{
Ahlam Mentag
}
